a lot of fixes including autocomplete and weather

+++ commands/utility/pookiedelete.js
@@ -1,4 +1,5 @@
 const { SlashCommandBuilder } = require('discord.js');
+const { Op } = require("sequelize");
 const { Pookiebears, UserPookies } = require('../../db/dbObjects.js');
 const fs = require('fs');
 
@@ -18,6 +19,8 @@ module.exports = {
 			const starryName = "starry night "+pookieName;
 			const starrySSR = "starry night "+pookieName+" SSR";
 			// equivalent to: DELETE from tags WHERE name = ?;
+			//loop according to length? maybe?
+			//const deletedPookie = await Pookiebears.findAll({ where: {pookie_name: pookieName}, rarity: { [Op.like]: '%+%',}});
 			const deletedPookie = await Pookiebears.findOne({ where: {pookie_name: pookieName}});
 			const deletedSSRPookie = await Pookiebears.findOne({ where: {pookie_name: ssrName}});
 			const deletedstarPookie = await Pookiebears.findOne({ where: {pookie_name: starryName}});
@@ -25,13 +28,17 @@ module.exports = {
 	
 			await UserPookies.destroy({ where: { pookie_id: deletedPookie.id } });
 			await UserPookies.destroy({ where: { pookie_id: deletedSSRPookie.id } });
+			if(deletedstarPookie)
 			await UserPookies.destroy({ where: { pookie_id: deletedstarPookie.id } });
+			if(deletedstarPookieSSR)
 			await UserPookies.destroy({ where: { pookie_id: deletedstarPookieSSR.id } });
 
 			//THIS FUCKING SUCKS SOO BAD ASDFKMASDFKSAF
 			const pookieDelete = await Pookiebears.destroy({ where: { pookie_name: pookieName } });
 			await Pookiebears.destroy({ where: { pookie_name: ssrName } });
+			if(deletedstarPookie)
 			await Pookiebears.destroy({ where: { pookie_name: starryName } });
+			if(deletedstarPookieSSR)
 			await Pookiebears.destroy({ where: { pookie_name: starrySSR } });
 
 			// Asynchronously delete a file
